[TSR-254] Persist the branches Hacienda already knows about
hacienda-history discovers an organization's branches, terminals and
consecutive counters by parsing the claves of everything the taxpayer has
emitted, and publishes them as SAVE_BRANCHES. Nothing consumed that event —
the topic had no subscriber — so discovery filled a queue that drained to a
DLQ. This adds the consumer.

Branches and terminals are ours, so the sync is deliberately additive:
ON CONFLICT DO NOTHING on (organization_id, code), which means an operator's
own name, type, address and phone survive every later sweep. Only the
consecutive counter moves, and only upwards — ON CONFLICT DO UPDATE with
GREATEST, which takes the same row write lock the sales allocator's
SELECT FOR UPDATE takes and is evaluated after waiting for it, so neither a
concurrent sale nor a reordered delivery can rewind a fiscal sequence.

The branch carries the organization's own contact details: the phone and the
provincia/cantón/distrito/barrio from its registered Hacienda profile, which is
the only place that data exists.

current_number becomes BIGINT. Hacienda's numeral ends in a ten-digit serial,
so the counter has to reach 9 999 999 999 and int4 stops at 2 147 483 647 —
importing any mature taxpayer's history would have overflowed on the first
write.

branches.type stops being Literal['stand','restaurant']. TSR-139 replaced those
two kinds with a per-org branch_types catalog and dropped the CHECK constraint,
but the request DTO kept the pair — so a discovered branch, which carries the
org's own first catalog code, could not be saved back through the same API it
came from. The type is now validated against that catalog instead of a frozen
list, which is what TSR-139 intended.

The service commits explicitly: the legacy context manager logs and swallows
commit failures, so without it a failed write would be acknowledged and the
message silently lost.

+++ app/models/consecutive.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 import uuid
 from typing import Optional
 
-from sqlalchemy import ForeignKey, Index, Integer, String, UniqueConstraint
+from sqlalchemy import BigInteger, ForeignKey, Index, Integer, String, UniqueConstraint
 from sqlalchemy.dialects.postgresql import UUID
 from sqlalchemy.orm import Mapped, mapped_column
 
@@ -27,7 +27,7 @@ class Consecutive(Base, TimestampMixin):
         ForeignKey("document_types.id", ondelete="RESTRICT"),
         nullable=False,
     )
-    current_number: Mapped[int] = mapped_column(Integer, nullable=False, default=0)
+    current_number: Mapped[int] = mapped_column(BigInteger, nullable=False, default=0)
     created_by: Mapped[str] = mapped_column(String(255), nullable=False)
 
     __table_args__ = (
